Use master as fallback for non-existent branches
When a PR is made directly from an upstream branch or to a different
branch than `master` or `stable`, the respective branch at the other two
repositories might not exist. So let's better check beforehand and
softly fallback to `master` otherwise. The same procedure is e.g. used
for CircleCi.