Add a Dockerfile to build an agent
This Dockerfile will allow administrator to quickly and easily deploy an agent.
The dependencies were extracted from ansible/roles/test_deps/tasks/main.yml.
A user can run the agent with:
$ docker build -t buildkite_agent buildkite
$ docker run -e BUILDKITE_AGENT_TOKEN='AGENT-TOKEN-HERE' buildkite_agent
Note that since this image is not intended to be distributed, the usual cleanup isn't done
(and no attempt was made to minimize the dependencies).