Buildkite: Remove a problematic git clean, use less generic build folder name
That git clean was causing an issue on a dedicated runner,
for no apparent reason. As it was there due to an abundance
of caution (many git clean are performed), we are safe to remove
this one.
Additionally, the build folder name was renamed to reduce likelihood
of conflict with any tested project's folder.